requirements.txt
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.240
huggingface_hub==1.2.3
pillow==12.0.0
pandas==2.3.3
tqdm==4.67.1
numpy==2.2.6
timm==1.0.22
boto3==1.42.16
tensorboard==2.20.0
tqdm==4.67.1
matplotlib==3.10.6
python-dotenv==1.0.1
